Basic functionality all working now!
For a given location and heading (direction of travel) will tell me
whether heads_i_lose (headwind) or tails_i_win (tailwind).

Since wind direction is the from direction and heading (as far as I'm
concerned) is the to direction, a head wind is defined as anything 2
compass points either side (on a 16-point compass). A tailwind is
anything else - which isn't true, but I didn't say I was finished yet.

Needs lots of tarting up as well.