Make decoder work back from step 11 to 3
Practically there. Adds in decoding steps 5 to 3. Not sure yet if
checking for last bit being 1 is the correct way to assess if a negative
number (steps 5 and 3), but can investigate that later.

There is also a lot of redundancy I'm sure i can remove. I.e. continual
conversion between binary numbers and the string representation of that
binary number. Some of this may not actually be necessary.