Bug fix - nth_wrap didn't work for last items in a list
I only noticed this just recently, which means it has been wrong for
about a year! I think it's just luck I'd never come across an issue with
it not working.

I don't know if I'd got half way through writing this and forgot to
finish it off, or whether I was just being an imbecile: If the remainder
is 0 then that should return the last item in a list, not just the
number 1.