Yay! Working prototype. Will sum up the distances of headwinds...
...for a polyline route.

I'd made silly mistakes like forgetting which order the list of tuples
were in. I.e {Distances, Wind} or {Wind, Distances}. It doesn't really
matter which order I use, but I should back sure it is consistent and
matches the variable naming.

Now need to extend this to sum up all wind directions and check for
other bugs (such as sum of distances making sense, headwinds are
actually headwinds and not tailwinds, etc).