Handle the reverse route
I don't want to have to query OSRM for th ereverse route - it's
pointless. Instead reverse the route we already have - because I don't
actually need a properly ordered reverse route, just a list of
directions and distances.

In order to do this made sense to re-jig some of the functions a bit:

- Need date and time at top level, not buried in get_weather since
  reverse route or normal route is also dependent on time of day
- Don't convert to distances and compass directions in one swoop since
  it is easier to reverse the heading angles rather than compass points